refactor(seeder): drive agent via shared generateLLMText, drop @anthropic-ai/sdk
The agent now calls @langfuse/shared's generateLLMText — the AI SDK v7
boundary that also backs the worker's LLM-as-a-judge — instead of a
second Anthropic SDK. generateLLMText is a single-call primitive (it
rejects tool execute fns by design), so the multi-turn loop stays here:
declare tools via createLLMToolSet, read toolCalls, run them, feed tool
messages back. The env key is wrapped into an Anthropic LLMConnection
(encrypt() with the Job's ENCRYPTION_KEY). Removes the @anthropic-ai/sdk
dependency and the knip workaround it needed.

Split into agent.ts (thin bootstrap, no src/ imports, always emits a
result block) + agent-main.ts (logic, imports the shared server barrel),
mirroring cli.ts/cli-main.ts so an env-schema failure still reports
cleanly. agent-main disconnects Prisma+Redis in a finally — the barrel
opens a Redis client whose reconnect loop would otherwise pin the event
loop and hang the Job to its deadline (verified: clean exit in ~5s).

Verified: typecheck, eslint --max-warnings 0, knip, build all clean;
runtime smoke test emits the marker block and exits 1 on the error path.
